General commenting and code cleanup to make vscode shut up

+++ src/entities.cpp
@@ -209,7 +209,7 @@ void checkitems()
     if(e.type==NOTUSED) continue;
     if(!ents[i].spawned && e.type!=TELEPORT && e.type!=JUMPPAD) continue;
     if(OUTBORD(e.x, e.y)) continue;
-    vec v = { e.x, e.y, S(e.x, e.y)->floor+player1->eyeheight };
+    vec v = { static_cast<float>(e.x), static_cast<float>(e.y), S(e.x, e.y)->floor+player1->eyeheight };
     vdist(dist, t, player1->o, v);
     if(dist<(e.type==TELEPORT ? 4 : 2.5)) pickup(i, player1);
   };
